Large rework due to a fundamental misunderstanding of how HTB works.
Switch to using a DRR below each host HTB classes instead of HTB
sub classes.  I had made the bad assumption that HTB divides bandwidth
along the class hierarchy. In reality what it does is only divide
bandwidth at the leaves according to the class priority. This meant
that in the previous version high or medium priority traffic from one host
would affect low priority traffic for another. This may be useful to
someone but isn't what I was aiming for.

Also removed the limitations around the number of hosts by making full
use of the major namespace. QDiscs are named and created in depth
first order.

Lots of small cleanups as well.